feat: Seperate admin page
